+++ qc_grader/challenges/fall_2022/lab4.py
@@ -11,7 +11,7 @@ from qiskit.visualization import *
 from qiskit_nature import settings
 
 from qc_grader.grader.grade import grade, get_problem_set, handle_submit_response
-from qc_grader.grader.common import circuit_to_json, optimizerresult_to_json
+from qc_grader.common.serializer import circuit_to_json, optimizerresult_to_json
 
 settings.dict_aux_operators = True
 settings.dict_aux_operators = True
