+++ qc_grader/challenges/programming_challenge_aus/problem_set.py
@@ -9,7 +9,7 @@ from qiskit import QuantumCircuit
 from qc_grader.grader.grade import (
     grade, get_problem_set
 )
-from qc_grader.grader.common import (
+from qc_grader.common.serializer import (
     circuit_to_json
 )
