+++ qc_grader/challenges/qgss_2022/lab2.py
@@ -12,7 +12,11 @@ from qiskit.opflow.primitive_ops import PauliOp
 from qiskit.opflow.primitive_ops.pauli_sum_op import PauliSumOp
 
 from qc_grader.grader.grade import grade
-from qc_grader.grader.common import circuit_to_json, pauliop_to_json, paulisumop_to_json
+from qc_grader.common.serializer import (
+    circuit_to_json,
+    pauliop_to_json,
+    paulisumop_to_json
+)
 
 
 _challenge_id = 'qgss_2022'
