+++ qc_grader/challenges/qgss_2023/lab2.py
@@ -1,5 +1,5 @@
 from typeguard import typechecked
-from typing import Dict, List, Sequence, Union
+from typing import Dict, Sequence, Union
 
 from qiskit import QuantumCircuit
 from qiskit.circuit import Parameter
@@ -8,7 +8,7 @@ from qiskit.quantum_info import SparsePauliOp
 from qiskit.primitives.base.sampler_result import SamplerResult
 
 from qc_grader.grader.grade import grade
-from qc_grader.grader.common import circuit_to_json, serialize_answer
+from qc_grader.common.serializer import circuit_to_json, serialize_answer
 
 
 _challenge_id = "qgss_2023"
