+++ qc_grader/common/serializer.py
@@ -11,43 +11,29 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-from fractions import Fraction
-from functools import wraps
-import json
-import logging
 import inspect
+import json
 
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
-import warnings
+from fractions import Fraction
+from typing import Any, Dict, List, Optional, Union
+
+from networkx.classes import Graph
 
 from qiskit import QuantumCircuit
-from qiskit_algorithms.optimizers import OptimizerResult
-from qiskit.circuit import Barrier, Gate, Instruction, Measure, Parameter
-from qiskit.circuit.library import UGate, U3Gate, CXGate, TwoLocal
-from qiskit.quantum_info.operators import Pauli as PauliOp
+from qiskit.circuit import Parameter
+from qiskit.circuit.library import TwoLocal
 from qiskit.primitives import SamplerResult, EstimatorResult
+from qiskit.qobj import PulseQobj, QasmQobj
+from qiskit.quantum_info import SparsePauliOp
+from qiskit.quantum_info.operators import Pauli as PauliOp
+from qiskit.result import ProbDistribution, QuasiDistribution
 from qiskit_aer.jobs import AerJob
 from qiskit_aer.noise import NoiseModel
-from qiskit_ibm_provider import IBMProvider, IBMProviderError
-from qiskit.quantum_info import SparsePauliOp
-from networkx.classes import Graph
-from qiskit_ibm_provider.job import IBMCircuitJob as IBMQJob
-from qiskit.qobj import PulseQobj, QasmQobj
-from qiskit.result import ProbDistribution, QuasiDistribution
 from qiskit_algorithms.minimum_eigensolvers.vqe import VQEResult
+from qiskit_algorithms.optimizers import OptimizerResult
+from qiskit_ibm_provider.job import IBMCircuitJob as IBMQJob
 
-from networkx import Graph
-
-
-ValidationResult = Tuple[bool, Optional[Union[str, int, float]]]
-
-
-class MaxContentError(BaseException):
-    def __init__(self, content_length: int, max_content_length: int) -> None:
-        self.message = f'Max content length ({max_content_length}) exceeded: {content_length}'
-
-    def __str__(self) -> str:
-        return self.message
+from qc_grader.common import get_job_urls
 
 
 class QObjEncoder(json.encoder.JSONEncoder):
@@ -68,6 +54,19 @@ class QObjEncoder(json.encoder.JSONEncoder):
             return {'__class__': 'Parameter', 'name': obj.name, 'uuid': str(obj._uuid)}
 
         return json.JSONEncoder.default(self, obj)
+
+
+def to_json(result: Any, skip: List[str] = []) -> str:
+    if result is None:
+        return ''
+    as_dict = {}
+    import inspect
+    for name, value in inspect.getmembers(result):
+        if not name.startswith('_') and name not in skip and \
+                not inspect.ismethod(value) and not inspect.isfunction(value):
+            as_dict[name] = value
+
+    return json.dumps(as_dict, cls=QObjEncoder)
 
 
 def circuit_to_dict(
@@ -135,6 +134,16 @@ def optimizerresult_to_json(
     return json.dumps(result, cls=QObjEncoder)
 
 
+def quasidistribution_to_json(
+    op: QuasiDistribution
+) -> str:
+    return json.dumps({
+        'data': str(op),
+        'shots': op.shots if hasattr(op, 'shots') else None,
+        'stddev_upper_bound': op.stddev_upper_bound if hasattr(op, 'stddev_upper_bound') else None
+    }, cls=QObjEncoder)
+
+
 def samplerresult_to_json(
     op: SamplerResult
 ) -> str:
@@ -162,15 +171,6 @@ def estimatorresult_to_json(
     }, cls=QObjEncoder)
 
 
-def quasidistribution_to_json(
-    op: QuasiDistribution
-) -> str:
-    return json.dumps({
-        'data': str(op),
-        'shots': op.shots if hasattr(op, 'shots') else None,
-        'stddev_upper_bound': op.stddev_upper_bound if hasattr(op, 'stddev_upper_bound') else None
-    }, cls=QObjEncoder)
-
 def probdistribution_to_json(
     op: ProbDistribution
 ) -> str:
@@ -178,7 +178,8 @@ def probdistribution_to_json(
         'data': str(op),
         'shots': op.shots,
     }, cls=QObjEncoder)
-    
+
+
 def vqeresult_to_json(
     result: VQEResult
 ) -> str:
@@ -186,156 +187,6 @@ def vqeresult_to_json(
         'eigenvalue': result.eigenvalue
         # TODO: figure out what other parmeters need to be included
     }, cls=QObjEncoder)
-
-
-def to_json(result: Any, skip: List[str] = []) -> str:
-    if result is None:
-        return ''
-    as_dict = {}
-    import inspect
-    for name, value in inspect.getmembers(result):
-        if not name.startswith('_') and name not in skip and \
-                not inspect.ismethod(value) and not inspect.isfunction(value):
-            as_dict[name] = value
-
-    return json.dumps(as_dict, cls=QObjEncoder)
-
-
-def cached(key_function: Callable) -> Callable:
-    def _decorator(f: Any) -> Callable:
-        f.__cache = {}
-        @wraps(f)
-        def _decorated(*args: Any, **kwargs: Any) -> int:
-            key = key_function(*args, **kwargs)
-            if key not in f.__cache:
-                f.__cache[key] = f(*args, **kwargs)
-            return f.__cache[key]
-        return _decorated
-    return _decorator
-
-
-def gate_key(gate: Gate) -> Tuple[str, int]:
-    return gate.name, gate.num_qubits
-
-
-@cached(gate_key)
-def gate_cost(gate: Gate) -> int:
-    if isinstance(gate, (UGate, U3Gate)):
-        return 1
-    elif isinstance(gate, CXGate):
-        return 10
-    elif isinstance(gate, (Measure, Barrier)):
-        return 0
-    return sum(map(gate_cost, (g for g, _, _ in gate.definition.data)))
-
-
-def compute_cost(circuit: Union[Instruction, QuantumCircuit]) -> int:
-    print('Computing cost...')
-    circuit_data = None
-    if isinstance(circuit, QuantumCircuit):
-        circuit_data = circuit.data
-    elif isinstance(circuit, Instruction):
-        circuit_data = circuit.definition.data
-    else:
-        raise Exception(f'Unable to obtain circuit data from {type(circuit)}')
-
-    return sum(map(gate_cost, (g for g, _, _ in circuit_data)))
-
-
-def uses_multiqubit_gate(circuit: QuantumCircuit) -> bool:
-    circuit_data = None
-    if isinstance(circuit, QuantumCircuit):
-        circuit_data = circuit.data
-    elif isinstance(circuit, Instruction) and circuit.definition is not None:
-        circuit_data = circuit.definition.data
-    else:
-        raise Exception(f'Unable to obtain circuit data from {type(circuit)}')
-
-    for g, _, _ in circuit_data:
-        if isinstance(g, (Barrier, Measure)):
-            continue
-        elif isinstance(g, Gate):
-            if g.num_qubits > 1:
-                return True
-        elif isinstance(g, (QuantumCircuit, Instruction)) and uses_multiqubit_gate(g):
-            return True
-
-    return False
-
-
-def calc_depth(qc: QuantumCircuit) -> Tuple[int, int]:
-    from qiskit.compiler import transpile
-
-    tqc = transpile(
-        qc, basis_gates=['u1', 'u2', 'u3', 'cx'], optimization_level=0
-    )
-    num_ops = tqc.count_ops()
-    depth = tqc.depth()
-    return depth, num_ops
-
-
-def get_provider(
-    hub: Optional[str] = None,
-    group: Optional[str] = None,
-    project: Optional[str] = None,
-    load_account_fallback: Optional[bool] = True
-) -> IBMProvider:
-    with warnings.catch_warnings():
-        warnings.simplefilter('ignore')
-
-        ibmq_logger = logging.getLogger('qiskit_ibm_provider')
-        current_level = ibmq_logger.level
-        ibmq_logger.setLevel(logging.ERROR)
-
-        provider = None
-
-        # get providers
-        if hub or group or project:
-            try:
-                provider = IBMProvider(instance=f'{hub}/{group}/{project}')
-            except IBMProviderError:
-                pass
-
-        # handle no correct provider found
-        if provider is None and load_account_fallback:
-            provider = IBMProvider()
-
-        ibmq_logger.setLevel(current_level)
-        return provider
-
-
-def get_job(
-    job_id: str,
-    hub: Optional[str] = None,
-    group: Optional[str] = None,
-    project: Optional[str] = None,
-    load_account_fallback: Optional[bool] = True
-) -> Optional[IBMQJob]:
-    try:
-        provider = get_provider(hub, group, project, load_account_fallback)
-        job = provider.backends.retrieve_job(job_id)
-        return job
-    except Exception:
-        pass
-
-    return None
-
-
-def get_job_urls(
-    job: Union[str, IBMQJob],
-    hub: Optional[str] = None,
-    group: Optional[str] = None,
-    project: Optional[str] = None,
-    load_account_fallback: Optional[bool] = True
-) -> Tuple[Optional[str], Optional[str]]:
-    try:
-        job_id = job.job_id() if isinstance(job, IBMQJob) else job
-        provider = get_provider(hub, group, project, load_account_fallback)
-        download_url = provider._api_client.account_api.job(job_id).download_url()['url']
-        result_url = provider._api_client.account_api.job(job_id).result_url()['url']
-        return download_url, result_url
-    except Exception:
-        return None, None
 
 
 def serialize_job(job: IBMQJob) -> Optional[Dict[str, str]]:
@@ -416,10 +267,3 @@ def serialize_answer(answer: Any, **kwargs: bool) -> Optional[str]:
         payload = json.dumps(answer, skipkeys=True, cls=QObjEncoder)
 
     return payload
-
-
-def normalize_slash(url: str) -> str:
-    if url[-1] != '/':
-        url += '/'
-
-    return url
