+++ qc_grader/grader/grade.py
@@ -18,6 +18,9 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from qiskit import transpile
 from qiskit_ibm_provider.job import IBMCircuitJob as IBMQJob
 
+from qc_grader.common import calc_depth, get_provider
+from qc_grader.common.serializer import serialize_answer
+
 from .api import (
     get_access_token,
     get_grading_endpoint,
@@ -25,12 +28,6 @@ from .api import (
     get_submission_endpoint,
     send_request,
     do_grade_only
-)
-from .common import (
-    ValidationResult,
-    calc_depth,
-    get_provider,
-    serialize_answer
 )
